Fix some compiler warnings
git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@9123 21252fe5-60e8-0310-8ff0-c1ae1001594e

+++ libraries/core/source/regfrmtrans.cpp
@@ -192,7 +192,7 @@ ATermAppl translate_reg_frms_appl(ATermAppl part)
     if (nr_args > 0)
     {
       MCRL2_SYSTEM_SPECIFIC_ALLOCA(args,ATerm,nr_args);
-      for (int i = 0; i < nr_args; i++)
+      for (size_t i = 0; i < nr_args; i++)
       {
         ATerm arg = ATgetArgument(part, i);
         if (ATgetType(arg) == AT_APPL)
